requirements.txt
groq
speechRecognition
pyttsx3
pyaudio
python-dotenv
streamlit

# pip install -r requirements.txt